Backend/requirements.txt
fastapi
uvicorn[standard]
requests
qrcode
pillow
python-dotenv
google-generativeai
beautifulsoup4